+++ frontend/src/api/trips/getTrips.ts
