+++ frontend/src/hooks/api/useTripsQuery.ts
@@ -3,7 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import type { TripsData } from '@type/trips';
 import type { AxiosError } from 'axios';
 
-import { getTrips } from '@api/trips/trips';
+import { getTrips } from '@api/trips/getTrips';
 
 export const useTripsQuery = () => {
   const { data: tripsData } = useQuery<TripsData[], AxiosError>(['trips'], getTrips, {
